+++ aws/services/gen_services.go
@@ -67,6 +67,8 @@ import (
 	eventbridgetypes "github.com/aws/aws-sdk-go-v2/service/eventbridge/types"
 	iam "github.com/aws/aws-sdk-go-v2/service/iam"
 	iamtypes "github.com/aws/aws-sdk-go-v2/service/iam/types"
+	kinesis "github.com/aws/aws-sdk-go-v2/service/kinesis"
+	kinesistypes "github.com/aws/aws-sdk-go-v2/service/kinesis/types"
 	kms "github.com/aws/aws-sdk-go-v2/service/kms"
 	kmstypes "github.com/aws/aws-sdk-go-v2/service/kms/types"
 	lambda "github.com/aws/aws-sdk-go-v2/service/lambda"
@@ -124,6 +126,7 @@ var ServiceNames = []string{
 	"stepfunctions",
 	"waf",
 	"configservice",
+	"kinesis",
 }
 
 var ResourceTypes = []string{
@@ -199,6 +202,7 @@ var ResourceTypes = []string{
 	"ipset",
 	"rulegroup",
 	"configrule",
+	"stream",
 }
 
 var ServicePerAPI = map[string]string{
@@ -235,6 +239,7 @@ var ServicePerAPI = map[string]string{
 	"sfn":                    "stepfunctions",
 	"wafv2":                  "waf",
 	"configservice":          "configservice",
+	"kinesis":                "kinesis",
 }
 
 var ServicePerResourceType = map[string]string{
@@ -310,6 +315,7 @@ var ServicePerResourceType = map[string]string{
 	"ipset":               "waf",
 	"rulegroup":           "waf",
 	"configrule":          "configservice",
+	"stream":              "kinesis",
 }
 
 var APIPerResourceType = map[string]string{
@@ -385,6 +391,7 @@ var APIPerResourceType = map[string]string{
 	"ipset":               "wafv2",
 	"rulegroup":           "wafv2",
 	"configrule":          "configservice",
+	"stream":              "kinesis",
 }
 
 type Infra struct {
@@ -4395,4 +4402,132 @@ func (s *Configservice) FetchByType(ctx context.Context, t string) (cloud.GraphA
 
 func (s *Configservice) IsSyncDisabled() bool {
 	return !getBool(s.config, "aws.configservice.sync", true)
+}
+
+type Kinesis struct {
+	fetcher         fetch.Fetcher
+	region, profile string
+	config          map[string]any
+	log             *logger.Logger
+	KinesisClient   *kinesis.Client
+}
+
+func NewKinesis(cfg aws.Config, profile string, extraConf map[string]any, log *logger.Logger) cloud.Service {
+	region := cfg.Region
+	kinesisClient := kinesis.NewFromConfig(cfg)
+
+	fetchConfig := awsfetch.NewConfig(
+		kinesisClient,
+	)
+	fetchConfig.Extra = extraConf
+	fetchConfig.Log = log
+
+	return &Kinesis{
+		KinesisClient: kinesisClient,
+		fetcher:       fetch.NewFetcher(awsfetch.BuildKinesisFetchFuncs(fetchConfig)),
+		config:        extraConf,
+		region:        region,
+		profile:       profile,
+		log:           log,
+	}
+}
+
+func (s *Kinesis) Name() string {
+	return "kinesis"
+}
+
+func (s *Kinesis) Region() string {
+	return s.region
+}
+
+func (s *Kinesis) Profile() string {
+	return s.profile
+}
+
+func (s *Kinesis) ResourceTypes() []string {
+	return []string{
+		"stream",
+	}
+}
+
+func (s *Kinesis) Fetch(ctx context.Context) (cloud.GraphAPI, error) {
+	if s.IsSyncDisabled() {
+		return graph.NewGraph(), nil
+	}
+
+	allErrors := new(fetch.Error)
+
+	gph, err := s.fetcher.Fetch(context.WithValue(ctx, "region", s.region))
+	defer s.fetcher.Reset()
+
+	for _, e := range *fetch.WrapError(err) {
+		switch ee := e.(type) {
+		case nil:
+			continue
+		default:
+			var ae smithy.APIError
+			if errors.As(ee, &ae) && ae.ErrorMessage() == accessDenied {
+				allErrors.Add(cloud.ErrFetchAccessDenied)
+			} else {
+				allErrors.Add(ee)
+			}
+		}
+	}
+
+	if err := gph.AddResource(graph.InitResource(cloud.Region, s.region)); err != nil {
+		return gph, err
+	}
+
+	snap := gph.AsRDFGraphSnaphot()
+
+	errc := make(chan error)
+	var wg sync.WaitGroup
+	if getBool(s.config, "aws.kinesis.stream.sync", true) {
+		list, err := s.fetcher.Get("stream_objects")
+		if err != nil {
+			return gph, err
+		}
+		if _, ok := list.([]kinesistypes.StreamSummary); !ok {
+			return gph, errors.New("cannot cast to '[]kinesistypes.StreamSummary' type from fetch context")
+		}
+		for _, r := range list.([]kinesistypes.StreamSummary) {
+			for _, fn := range addParentsFns["stream"] {
+				wg.Add(1)
+				go func(f addParentFn, snap tstore.RDFGraph, region string, res *kinesistypes.StreamSummary) {
+					defer wg.Done()
+					err := f(gph, snap, region, res)
+					if err != nil {
+						errc <- err
+						return
+					}
+				}(fn, snap, s.region, &r)
+			}
+		}
+	}
+
+	go func() {
+		wg.Wait()
+		close(errc)
+	}()
+
+	for err := range errc {
+		if err != nil {
+			allErrors.Add(err)
+		}
+	}
+
+	if allErrors.Any() {
+		return gph, allErrors
+	}
+
+	return gph, nil
+}
+
+func (s *Kinesis) FetchByType(ctx context.Context, t string) (cloud.GraphAPI, error) {
+	defer s.fetcher.Reset()
+	return s.fetcher.FetchByType(context.WithValue(ctx, "region", s.region), t)
+}
+
+func (s *Kinesis) IsSyncDisabled() bool {
+	return !getBool(s.config, "aws.kinesis.sync", true)
 }