+++ aws/services/init.go
@@ -29,7 +29,7 @@ import (
 var (
 	AccessService, InfraService, StorageService, MessagingService, DNSService, LambdaService, MonitoringService, CdnService, CloudformationService cloud.Service
 	EksService, DynamodbService, SecretsmanagerService, ApigatewayService, SsmService, EfsService, CloudtrailService, CloudwatchlogsService        cloud.Service
-	ElasticacheService, EventbridgeService, StepfunctionsService, WafService, ConfigserviceService                                                 cloud.Service
+	ElasticacheService, EventbridgeService, StepfunctionsService, WafService, ConfigserviceService, KinesisService                                 cloud.Service
 )
 
 func Init(profile, region string, extraConf map[string]any, log *logger.Logger, profileSetterCallback func(val string) error, enableNetworkMonitor bool) error {
@@ -67,6 +67,7 @@ func Init(profile, region string, extraConf map[string]any, log *logger.Logger, 
 	StepfunctionsService = NewStepfunctions(cfg, profile, extraConf, log)
 	WafService = NewWaf(cfg, profile, extraConf, log)
 	ConfigserviceService = NewConfigservice(cfg, profile, extraConf, log)
+	KinesisService = NewKinesis(cfg, profile, extraConf, log)
 
 	cloud.ServiceRegistry[InfraService.Name()] = InfraService
 	cloud.ServiceRegistry[AccessService.Name()] = AccessService
@@ -90,6 +91,7 @@ func Init(profile, region string, extraConf map[string]any, log *logger.Logger, 
 	cloud.ServiceRegistry[StepfunctionsService.Name()] = StepfunctionsService
 	cloud.ServiceRegistry[WafService.Name()] = WafService
 	cloud.ServiceRegistry[ConfigserviceService.Name()] = ConfigserviceService
+	cloud.ServiceRegistry[KinesisService.Name()] = KinesisService
 
 	awsspec.CommandFactory = &awsspec.AWSFactory{
 		Log: log,